requirements-lock.txt
boto3==1.28.76
botocore==1.31.76
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
